+++ example/playground/maleo.config.js
@@ -3,6 +3,8 @@ const cssPlugin = require('@airy/maleo-css-plugin');
 
 module.exports = tsPlugin(
   cssPlugin({
-    env: 'production',
+    cssPluginOptions: {
+      enableISL: true
+    }
   }),
 );