+++ packages/plugins/css-plugin/CHANGELOG.md
@@ -3,7 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0](https://github.com/alvinkl/maleo.js/compare/@airy/maleo-css-plugin@0.0.12-canary.0...@airy/maleo-css-plugin@0.1.0) (2019-03-26)
+
+
+### Features
+
+* **maleo-plugin:** improve maleo plugin [skip ci] ([#154](https://github.com/alvinkl/maleo.js/issues/154)) ([10638c5](https://github.com/alvinkl/maleo.js/commit/10638c5)), closes [#152](https://github.com/alvinkl/maleo.js/issues/152)
+
 ## [0.0.13-canary.0](https://github.com/airyrooms/maleo.js/compare/@airy/maleo-css-plugin@0.0.12-canary.0...@airy/maleo-css-plugin@0.0.13-canary.0) (2019-03-25)
+
+
+
+
+
+## [0.0.13](https://github.com/alvinkl/maleo.js/compare/@airy/maleo-css-plugin@0.0.12-canary.0...@airy/maleo-css-plugin@0.0.13) (2019-03-25)
 
 **Note:** Version bump only for package @airy/maleo-css-plugin